\name{sapscaleR-package}
\alias{sapscaleR-package}
\alias{sapscaleR}
\docType{package}
\title{sapscaleR}
\description{
Models for up-scaling sap flux density measurements to whole-tree volumetric
sap flow rates using circular and irregular-trunk geometry.
}
\details{
The package implements whole-trunk sap flow estimators described by Burnett
et al. (2026) for circular, Steiner-corrected, and corrected irregular trunk
cross-sections.
}
\author{
Michael W. Burnett \email{mburnett@ucsb.edu}
}
